% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_fis_broker_url.R
\name{compose_fis_broker_url}
\alias{compose_fis_broker_url}
\title{Compose FIS-Broker URL}
\usage{
compose_fis_broker_url(
  cmd = "user_login",
  session_id = NULL,
  type = NULL,
  id = NULL
)
}
\arguments{
\item{cmd}{command (default: "user login")}

\item{session_id}{as retrieved by \code{\link{login_to_fis_broker}},
(default: NULL)}

\item{type}{dataset type, i.e. "ATOM", "WFS" or "WMS" (default: NULL)}

\item{id}{dataset id (default: NULL)}
}
\value{
composed FIS-Broker URL
}
\description{
Compose FIS-Broker URL
}
