% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{create_info_page_hrefs}
\alias{create_info_page_hrefs}
\title{Create Info Page HREFS}
\usage{
create_info_page_hrefs(overview, method = 2L, dbg = TRUE)
}
\arguments{
\item{overview}{overview tibble as retrieved by \code{\link{get_dataset_overview}}}

\item{method}{method (either 1 or 2) for creating hrefs (default: 1)}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
vector with hrefs
}
\description{
Create Info Page HREFS
}
