% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_overview.R
\name{get_dataset_overview}
\alias{get_dataset_overview}
\title{Get Dataset Overview}
\usage{
get_dataset_overview(dbg = TRUE)
}
\arguments{
\item{dbg}{whether or not to show debug messages}
}
\value{
tibble with 7 columns and rows equal to number of datasets (
one for each dataset type, e.g WMS, WFS, ATOM)
\describe{
\item{category_id}{Category id of dataset (artifically generated)}
\item{category_name}{Category name of dataset)}
\item{dataset_id}{Dataset id, artifically generated!!!}
\item{dataset_name}{Dataset name as listed on table}
\item{dataset_name_href}{Link of dataset name. JS session Id was replaced
with a placeholder}
\item{dataset_type}{Dataset type as listed on table one of \code{ATOM}, \code{WFS} or
\code{WMS}}
\item{dataset_type_href}{Link of dataset type. JS session Id was replaced
with a placeholder}
}
}
\description{
Get Dataset Overview
}
\examples{
fb_dataset_overview <- kwb.fisbroker::get_dataset_overview()
}
