% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_urls}
\alias{get_urls}
\title{Get Urls}
\usage{
get_urls(..., key. = NULL)
}
\arguments{
\item{\dots}{further arguments passed to \code{\link[kwb.utils]{resolve}}}

\item{key.}{optional. Name of element to be returned from the URL dictionary.
If \code{NULL} (the default), the whole dictionary (a list) is returned.}
}
\value{
return \code{base} and \code{wfs} urls of FIS-Broker
}
\description{
Get Urls
}
\examples{
# List of all defined (partial) URls
get_urls()

# One specific URL with placeholders replaced as given
get_urls(
  key. = "href_type", sid = "my-session-id", type = "WFS", id = "my-id"
)
}
