% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wfs_data.R
\name{get_wfs_dataset}
\alias{get_wfs_dataset}
\title{Get WFS Dataset}
\usage{
get_wfs_dataset(
  dataset_id = "s_wfs_alkis_bezirk",
  service_version = "2.0.0",
  srs_name = "EPSG:4258",
  debug = TRUE
)
}
\arguments{
\item{dataset_id}{id of wfs dataset (default: "s_wfs_alkis_bezirk"). Needs to
checked out manually as described below:
\enumerate{
\item Go to: \url{https://fbinter.stadt-berlin.de/fb/}
\item Click on a WFS Dataset
\item Read the content of \code{Rechneraddresse} in the opened tab. The basename
of the url is the required dataset id!
}}

\item{service_version}{one of "1.0.0", "1.1.0" or "2.0.0" (default: "2.0.0")}

\item{srs_name}{one of "EPSG:4258" or "EPSG:25833" (default: "EPSG:4258")}

\item{debug}{prints debug messages if TRUE (default: TRUE)}
}
\value{
imports selected WFS dataset into R
}
\description{
Get WFS Dataset
}
\examples{
berlin_bezirke <- kwb.fisbroker::get_wfs_dataset(dataset_id = "s_wfs_alkis_bezirk")
}
\seealso{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_wfs_alkis_bezirk@senstadt&type=WFS}
}
