% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{httr_get_or_fail}
\alias{httr_get_or_fail}
\title{GET Request with Check for Error}
\usage{
httr_get_or_fail(url, handle = NULL)
}
\arguments{
\item{url}{URL to which to send a GET request}

\item{handle}{passed to \code{\link[httr]{GET}}}
}
\value{
If the status was not 200, the function returns what
\code{\link[httr]{GET}} returned
}
\description{
The function stops with error if the GET request returns status != 200
}
