% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_to_fis_broker.R
\name{login_to_fis_broker}
\alias{login_to_fis_broker}
\title{Login to FIS-Broker}
\usage{
login_to_fis_broker(dbg = TRUE)
}
\arguments{
\item{dbg}{print debug messages (default: TRUE)}
}
\value{
session id
}
\description{
Login to FIS-Broker
}
\examples{
login_to_fis_broker()

}
