% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{parse_for_metadata}
\alias{parse_for_metadata}
\title{Parse HTML Tree for Metadata}
\usage{
parse_for_metadata(x, dbg = TRUE)
}
\arguments{
\item{x}{object of class "xml_node"}

\item{dbg}{whether or not to show debug messages}
}
\description{
Parse HTML Tree for Metadata
}
