% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_all_metadata}
\alias{read_all_metadata}
\title{Read All Metadata}
\usage{
read_all_metadata(overview = get_dataset_overview(), dbg = TRUE, method = 1L)
}
\arguments{
\item{overview}{overview tibble as retrieved by \code{\link{get_dataset_overview}},
(default: \code{\link{get_dataset_overview}})}

\item{dbg}{print debug messages (default: TRUE)}

\item{method}{method (either 1 or 2) used by \code{\link{create_info_page_hrefs}}
for creating hrefs (default: 1)}
}
\value{
tibble with all metdata information
}
\description{
Read All Metadata
}
\examples{
\dontrun{
overview <- kwb.fisbroker::get_dataset_overview()
overview_atom <- overview[overview$type == "ATOM",]
metadata_atom <- kwb.fisbroker::read_all_metadata(overview_atom)
overview_wfs <- overview[overview$type == "WFS",]
metadata_wfs <- kwb.fisbroker::read_all_metadata(overview_wfs)
overview_wms <- overview[overview$type == "WMS",]
metadata_wms <- kwb.fisbroker::read_all_metadata(overview_wms)
}
}
