% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(dataset_id = "s_wfs_alkis_bezirk", service_type = "WFS")
}
\arguments{
\item{dataset_id}{id of wfs dataset (default: "s_wfs_alkis_bezirk"). Needs to
checked out manually as described below:
\enumerate{
\item Go to: \url{https://fbinter.stadt-berlin.de/fb/}
\item Click on a "WFS" or "WMS" Dataset
\item Read the content of \code{Rechneraddresse} in the opened tab. The basename
of the url is the required dataset id!
}}

\item{service_type}{either "WFS" or "WMS" (default: "WFS")}
}
\value{
tibble with metadata for provided dataset_id
}
\description{
Read metadata
}
\examples{
berlin_bezirke_metadata <- read_metadata(dataset_id = "s_wfs_alkis_bezirk")
berlin_bezirke_metadata

}
\seealso{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_wfs_alkis_bezirk@senstadt&type=WFS}
}
