% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wfs.R
\name{read_wfs}
\alias{read_wfs}
\title{Read WFS Dataset from FIS-Broker}
\usage{
read_wfs(
  dataset_id = "s_wfs_alkis_bezirk",
  service_version = "2.0.0",
  srs = "EPSG:25833",
  encoding = "UTF-8",
  dbg = TRUE
)
}
\arguments{
\item{dataset_id}{id of wfs dataset (default: "s_wfs_alkis_bezirk"). Needs to
checked out manually as described below:
\enumerate{
\item Go to: \url{https://fbinter.stadt-berlin.de/fb/}
\item Click on a WFS Dataset
\item Read the content of \code{Rechneraddresse} in the opened tab. The basename
of the url is the required dataset id!
}}

\item{service_version}{one of "1.0.0", "1.1.0" or "2.0.0" (default: "2.0.0")}

\item{srs}{one of "EPSG:4258" or "EPSG:25833" (default: "EPSG:25833")}

\item{encoding}{default: UTF-8}

\item{dbg}{prints debug messages if TRUE (default: TRUE)}
}
\value{
imports selected WFS dataset into R
}
\description{
Read WFS Dataset from FIS-Broker
}
\examples{
berlin_bezirke <- kwb.fisbroker::read_wfs(dataset_id = "s_wfs_alkis_bezirk")
}
\seealso{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_wfs_alkis_bezirk@senstadt&type=WFS}
}
