% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{table_to_data_frame}
\alias{table_to_data_frame}
\title{HTML Table to Data Frame}
\usage{
table_to_data_frame(table_node, dbg = TRUE)
}
\arguments{
\item{dbg}{whether or not to show debug messages}

\item{x}{}
}
\description{
HTML Table to Data Frame
}
