% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_to}
\alias{write_to}
\title{Write to JSON or CSV}
\usage{
write_to(format, df)
}
\arguments{
\item{format}{either 'json' or 'csv'}

\item{df}{data frame}
}
\value{
writes provided df to working directory with same name
}
\description{
Write to JSON or CSV
}
