% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{create_info_page_hrefs}
\alias{create_info_page_hrefs}
\title{Create Full URLs to Info Pages about Datasets in Overview Table}
\usage{
create_info_page_hrefs(overview, dbg = TRUE)
}
\arguments{
\item{overview}{overview tibble as retrieved by \code{\link{get_dataset_overview}},
(default: \code{\link{get_dataset_overview}})}

\item{dbg}{whether or not to show debug messages (default: TRUE)}
}
\value{
vector with hrefs
}
\description{
Create Full URLs to Info Pages about Datasets in Overview Table
}
