% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_urls}
\alias{get_urls}
\title{Get Urls}
\usage{
get_urls(...)
}
\arguments{
\item{\dots}{further arguments passed to \code{\link[kwb.utils]{resolve}}}
}
\value{
return \code{base} and \code{wfs} urls of FIS-Broker
}
\description{
Get Urls
}
\examples{
fisbroker_urls <- get_urls()
fisbroker_urls

}
