% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dataset_id.R
\name{lookup_dataset_id}
\alias{lookup_dataset_id}
\title{Lookup ID Required for Download}
\usage{
lookup_dataset_id(overview, dataset_id, type)
}
\arguments{
\item{overview}{data frame as returned by
\code{kwb.fisbroker:::get_dataset_overview}}

\item{type}{dataset type, i.e. "ATOM", "WFS" or "WMS"}
}
\description{
Lookup ID Required for Download
}
