% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_all_metadata}
\alias{read_all_metadata}
\title{Read Metadata based on Overview on FIS Broker Datasets}
\usage{
read_all_metadata(overview = get_dataset_overview(), dbg = TRUE)
}
\arguments{
\item{overview}{data frame as returned by
\code{kwb.fisbroker:::get_dataset_overview}}

\item{dbg}{whether or not to show debug messages}
}
\description{
Read Metadata based on Overview on FIS Broker Datasets
}
